requirements.txt
torch>=2.0
torchvision>=0.15
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
Pillow>=9.0
wandb>=0.16
python-dotenv>=1.0
jupyter>=1.0
pytest>=7.0
ruff>=0.4
